refactor(database): 重构数据库模型及模式定义
- 将 Participant、PersonSign、Transcription 模型合并到数据库包中
- 调整 Participant 模型为简单表并添加创建时间字段，采用 UTC 时间
- 修正 User 模型主键为 BigInteger 类型
- 优化 Meeting 模型中导入的依赖并移除多余代码
- 扩展 schemas 模块，补充 Participant、PersonSign 等模式的定义
- 更新 User 模式中性别、用户角色和状态的校验规则，改用具体枚举值验证
- 调整部分字段的约束和默认值，如字符串最大长度和默认状态字段

+++ models/database/user.py
@@ -14,7 +14,7 @@ class User(AuditedBase):
     __tablename__ = "users"
 
     # 主键字段（AuditedBase 不包含 id，需显式定义）
-    id: Mapped[int] = mapped_column(
+    id: Mapped[BigInteger] = mapped_column(
         BigInteger, primary_key=True, autoincrement=True, comment="用户主键ID（自增）"
     )
